requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pdfplumber==0.11.4
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2024.1
sqlparse==0.5.1
tzdata==2024.1